DataSrc: load_data takes dtype, default float
load_data takes dtype argument which has the default set to float
so that it mirrors the previous behavior.

This is related to numpy.genfromtxt.

EUWorldDataSrc updated to set dtype to int as part of its
_fetchconv_postproc.

+++ datasrc.py
@@ -133,14 +133,14 @@ class DataSrc:
         raise ImportError("DataSrc:_load_hdr: No header found")
 
 
-    def load_data(self, fileName=None, delimiter=None, skip_header=None, converters=None, iHdrLine=None, usecols=None):
+    def load_data(self, fileName=None, dtype=float, delimiter=None, skip_header=None, converters=None, iHdrLine=None, usecols=None):
         """
             iHdrLine: the column header line among the skip_header lines, starts at 0
             """
         if fileName == None:
             fileName = self.localFileName
         print("INFO:DataSrc:Loading:{}".format(fileName))
-        self.data = numpy.genfromtxt(fileName, delimiter=delimiter, skip_header=skip_header, converters=converters, usecols=usecols)
+        self.data = numpy.genfromtxt(fileName, dtype = dtype, delimiter=delimiter, skip_header=skip_header, converters=converters, usecols=usecols)
         if (skip_header != None) and (iHdrLine != None):
             if (iHdrLine < skip_header):
                 self.hdr = self._load_hdr(fileName, delimiter, iHdrLine)
@@ -238,7 +238,7 @@ class EUWorldDataSrc(DataSrc):
         cols = tuple(cols)
         converters = { 0: lambda x: self.conv_date(x), cols[2]: lambda x: self.conv_geoid(x) }
         print(cols)
-        super().load_data(fileName=fileName, delimiter=",", skip_header=1, converters=converters, iHdrLine=0, usecols=cols)
+        super().load_data(fileName=fileName, dtype=int, delimiter=",", skip_header=1, converters=converters, iHdrLine=0, usecols=cols)
         print(self.data)
         input("DBG: CHeck postproc data")
